LLM_Judge_Experiments/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5
pydantic>=2.0.0
pillow>=9.0.0
tqdm>=4.0.0
python-dotenv>=1.0.0
